Removed MemInterface and all derived classes; merged all functionality into MMU.
