+++ src/lib/memory/MMU.cc
@@ -4,8 +4,73 @@
 namespace simeng {
 namespace memory {
 
-MMU::MMU(std::shared_ptr<Mem> memory, VAddrTranslator fn, uint64_t tid)
-    : memory_(memory), translate_(fn), tid_(tid) {}
+MMU::MMU(std::shared_ptr<Mem> memory, uint16_t latency, VAddrTranslator fn,
+         uint64_t tid)
+    : memory_(memory), latency_(latency), tid_(tid), translate_(fn) {}
+
+void MMU::tick() {
+  tickCounter_++;
+
+  while (pendingRequests_.size() > 0) {
+    const auto& request = pendingRequests_.front();
+
+    if (request.readyAt > tickCounter_) {
+      // Head of queue isn't ready yet; end cycle
+      break;
+    }
+
+    const auto& target = request.target;
+    uint64_t requestId = request.requestId;
+
+    if (request.write) {
+      const char* wdata = request.data.getAsVector<char>();
+      std::vector<char> dt(wdata, wdata + target.size);
+      // Responses to write requests are ignored by passing in a nullptr
+      // callback because they don't contain any information relevant to the
+      // simulation.
+      bufferRequest(memory::DataPacket(target.address, target.size,
+                                       memory::WRITE_REQUEST, requestId, dt),
+                    nullptr);
+    } else {
+      // Instantiate a callback function which will be invoked with the response
+      // to a read request.
+      auto fn = [this, target,
+                 requestId](struct memory::DataPacket packet) -> void {
+        if (packet.inFault_) {
+          completedReads_.push_back({target, RegisterValue(), requestId});
+          return;
+        }
+        completedReads_.push_back(
+            {target, RegisterValue(packet.data_.data(), packet.size_),
+             requestId});
+      };
+      bufferRequest(memory::DataPacket(target.address, target.size,
+                                       memory::READ_REQUEST, requestId),
+                    fn);
+    }
+
+    // Remove the request from the queue
+    pendingRequests_.pop();
+  }
+}
+
+void MMU::requestRead(const MemoryAccessTarget& target, uint64_t requestId) {
+  pendingRequests_.push({target, tickCounter_ + latency_, requestId});
+}
+
+void MMU::requestWrite(const MemoryAccessTarget& target,
+                       const RegisterValue& data) {
+  pendingRequests_.push({target, data, tickCounter_ + latency_});
+}
+
+const span<MemoryReadResult> MMU::getCompletedReads() const {
+  return {const_cast<MemoryReadResult*>(completedReads_.data()),
+          completedReads_.size()};
+}
+
+void MMU::clearCompletedReads() { completedReads_.clear(); }
+
+bool MMU::hasPendingRequests() const { return !pendingRequests_.empty(); }
 
 void MMU::bufferRequest(DataPacket request,
                         sendResponseToMemInterface sendResponse) {
@@ -29,6 +94,31 @@ void MMU::bufferRequest(DataPacket request,
 }
 
 void MMU::setTid(uint64_t tid) { tid_ = tid; }
+
+void MMU::requestInstrRead(const MemoryAccessTarget& target,
+                           uint64_t requestId) {
+  // Instantiate a callback function which will be invoked with the response
+  // to a read request.
+  auto fn = [this, target, requestId](memory::DataPacket dpkt) -> void {
+    if (dpkt.inFault_) {
+      completedInstrReads_.push_back({target, RegisterValue(), requestId});
+      return;
+    }
+    completedInstrReads_.push_back(
+        {target, RegisterValue(dpkt.data_.data(), dpkt.size_), requestId});
+  };
+
+  bufferRequest(memory::DataPacket(target.address, target.size,
+                                   memory::READ_REQUEST, requestId),
+                fn);
+}
+
+const span<MemoryReadResult> MMU::getCompletedInstrReads() const {
+  return {const_cast<MemoryReadResult*>(completedInstrReads_.data()),
+          completedInstrReads_.size()};
+}
+
+void MMU::clearCompletedIntrReads() { completedInstrReads_.clear(); }
 
 }  // namespace memory
 }  // namespace simeng
